chore: coordinate utils split out into own script
